feat(mobile): expo v54
