+++ apps/extension/package.json
@@ -143,11 +143,11 @@
     "prismjs": "1.30.0",
     "promise-memoize": "1.2.1",
     "punycode": "2.3.1",
-    "react": "19.0.0",
+    "react": "19.1.0",
     "react-animate-height": "3.1.1",
     "react-async-hook": "4.0.0",
     "react-call": "1.8.1",
-    "react-dom": "19.0.0",
+    "react-dom": "19.1.0",
     "react-dom-confetti": "0.2.0",
     "react-head": "3.4.2",
     "react-hook-form": "7.62.0",
@@ -218,8 +218,8 @@
     "@types/prismjs": "1.26.5",
     "@types/promise-memoize": "1.2.4",
     "@types/punycode": "2.1.4",
-    "@types/react": "19.0.12",
-    "@types/react-dom": "19.0.4",
+    "@types/react": "19.1.17",
+    "@types/react-dom": "19.1.11",
     "@types/react-router": "5.1.20",
     "@types/remote-redux-devtools": "0.5.8",
     "@types/valid-url": "1.0.7",
