+++ apps/mobile/babel.config.cjs
@@ -1,11 +1,6 @@
 module.exports = function (api) {
   api.cache(false);
   return {
-    plugins: [
-      '@lingui/babel-plugin-lingui-macro',
-      '@babel/plugin-transform-class-static-block',
-      'react-native-worklets/plugin',
-    ],
-    presets: ['babel-preset-expo'],
+    plugins: ['@lingui/babel-plugin-lingui-macro', 'react-native-worklets/plugin'],
   };
 };
