+++ apps/mobile/metro.config.cjs
@@ -2,7 +2,6 @@
 /* eslint-disable no-console */
 
 const path = require('path');
-const MetroSymlinksResolver = require('@rnx-kit/metro-resolver-symlinks');
 const { getSentryExpoConfig } = require('@sentry/react-native/metro');
 
 const projectRoot = __dirname;
@@ -20,34 +19,33 @@ config.resolver = {
   ...config.resolver,
   assetExts: config.resolver.assetExts.filter(ext => ext !== 'svg'),
   sourceExts: [...config.resolver.sourceExts, 'svg'],
-  unstable_enablePackageExports: true,
-  unstable_conditionNames: ['require', 'node', 'import'],
 };
 
 config.resolver.extraNodeModules = {
   stream: require.resolve('readable-stream'),
-  tslib: path.resolve(__dirname, 'node_modules/tslib'),
 };
 
 // #1 - Watch all files in the monorepo
 config.watchFolders = [workspaceRoot];
-// #2 - Force resolving nested modules to the folders below
-config.resolver.disableHierarchicalLookup = true;
-// #3 - Try resolving with project modules first, then workspace modules
+// #2 - Try resolving with project modules first, then workspace modules
 config.resolver.nodeModulesPaths = [
   path.resolve(projectRoot, 'node_modules'),
   path.resolve(workspaceRoot, 'node_modules'),
-  path.resolve(workspaceRoot, 'node_modules', '.pnpm', 'node_modules'),
 ];
-const symlinkResolver = MetroSymlinksResolver({
-  experimental_retryResolvingFromDisk: 'force',
-});
 
 config.resolver.resolveRequest = (context, moduleName, platform) => {
-  // TODO: either read tsconfig automatically or figure out another way of resolving tsconfig aliases.
-  // If we add more aliases in tsconfig, we need to add those to this if statement.
-  if (moduleName.startsWith('@/')) {
-    return context.resolveRequest(context, moduleName, platform);
+  // Handle @noble/hashes internal imports that aren't in exports field
+  // These work via file-based resolution but SDK 54 warns about them
+  if (moduleName === '@noble/hashes/crypto' || moduleName === '@noble/hashes/cryptoBrowser') {
+    // Use file-based resolution without warning
+    return context.resolveRequest(
+      {
+        ...context,
+        unstable_enablePackageExports: false,
+      },
+      moduleName,
+      platform
+    );
   }
 
   // Block @stacks/connect-ui which contains webpack-specific dynamic imports
@@ -71,14 +69,6 @@ config.resolver.resolveRequest = (context, moduleName, platform) => {
     );
   }
 
-  // Ensures resolution of the browser version of `@noble/hashes`s exports.
-  // Without this the node export is resolved, resulting in
-  // `crypto.getRandomValues` exceptions
-  // https://github.com/paulmillr/noble-hashes/blob/main/package.json#L47-L50
-  if (moduleName === '@noble/hashes/crypto') {
-    return context.resolveRequest(context, moduleName, platform);
-  }
-
   // Enable fast refresh for the UI package
   if (isDevelopment) {
     if (moduleName === '@leather.io/ui/native') {
@@ -99,14 +89,7 @@ config.resolver.resolveRequest = (context, moduleName, platform) => {
     }
   }
 
-  return symlinkResolver(context, moduleName, platform);
+  return context.resolveRequest(context, moduleName, platform);
 };
-
-config.transformer.getTransformOptions = async () => ({
-  transform: {
-    experimentalImportSupport: false,
-    inlineRequires: true,
-  },
-});
 
 module.exports = config;
