+++ apps/mobile/package.json
@@ -36,7 +36,7 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@expo/metro-config": "0.20.18",
+    "@expo/metro-config": "54.0.14",
     "@formatjs/intl-getcanonicallocales": "2.5.5",
     "@formatjs/intl-locale": "4.2.11",
     "@formatjs/intl-numberformat": "8.15.4",
@@ -63,17 +63,16 @@
     "@lingui/core": "5.4.0",
     "@lingui/react": "5.4.0",
     "@noble/hashes": "1.5.0",
-    "@react-native-async-storage/async-storage": "2.1.2",
-    "@react-native-firebase/app": "22.2.0",
-    "@react-native-firebase/messaging": "22.2.0",
+    "@react-native-async-storage/async-storage": "2.2.0",
+    "@react-native-firebase/app": "23.8.4",
+    "@react-native-firebase/messaging": "23.8.4",
     "@react-native/assets-registry": "0.73.1",
     "@react-navigation/native": "7.1.9",
     "@react-navigation/stack": "7.3.2",
     "@reduxjs/toolkit": "2.8.1",
-    "@rnx-kit/metro-resolver-symlinks": "0.2.3",
     "@scure/btc-signer": "1.6.0",
     "@sentry/cli": "2.42.2",
-    "@sentry/react-native": "6.14.0",
+    "@sentry/react-native": "7.2.0",
     "@shopify/flash-list": "2.0.2",
     "@shopify/restyle": "2.4.2",
     "@stacks/common": "7.0.2",
@@ -82,70 +81,70 @@
     "@stacks/transactions": "7.0.5",
     "@tanstack/react-query": "5.85.5",
     "@types/lodash.get": "4.4.9",
-    "@types/react-dom": "19.0.4",
+    "@types/react-dom": "19.1.11",
     "axios": "1.12.0",
     "bignumber.js": "9.1.2",
     "bitcoinjs-lib": "6.1.5",
     "buffer": "6.0.3",
     "dayjs": "1.11.13",
     "eas-build-cache-provider": "16.4.2",
-    "expo": "53.0.25",
+    "expo": "54.0.32",
     "expo-alternate-app-icons": "7.0.0",
-    "expo-application": "6.1.5",
-    "expo-asset": "11.1.7",
-    "expo-blur": "14.1.5",
-    "expo-build-properties": "0.14.8",
-    "expo-camera": "16.1.11",
-    "expo-clipboard": "7.1.5",
-    "expo-constants": "17.1.8",
-    "expo-crypto": "14.1.5",
-    "expo-dev-client": "5.2.4",
-    "expo-file-system": "18.1.11",
-    "expo-font": "13.3.2",
-    "expo-haptics": "14.1.4",
-    "expo-image": "2.4.1",
-    "expo-linear-gradient": "14.1.5",
-    "expo-linking": "7.1.7",
-    "expo-local-authentication": "16.0.5",
-    "expo-modules-autolinking": "2.1.10",
-    "expo-modules-core": "2.5.0",
-    "expo-notifications": "0.31.4",
-    "expo-router": "5.1.10",
-    "expo-secure-store": "14.2.3",
-    "expo-sharing": "13.1.5",
-    "expo-splash-screen": "0.30.10",
+    "expo-application": "7.0.8",
+    "expo-asset": "12.0.12",
+    "expo-blur": "15.0.8",
+    "expo-build-properties": "1.0.10",
+    "expo-camera": "17.0.10",
+    "expo-clipboard": "8.0.8",
+    "expo-constants": "18.0.13",
+    "expo-crypto": "15.0.8",
+    "expo-dev-client": "6.0.20",
+    "expo-file-system": "19.0.21",
+    "expo-font": "14.0.11",
+    "expo-haptics": "15.0.8",
+    "expo-image": "3.0.11",
+    "expo-linear-gradient": "15.0.8",
+    "expo-linking": "8.0.11",
+    "expo-local-authentication": "17.0.8",
+    "expo-modules-autolinking": "3.0.24",
+    "expo-modules-core": "3.0.29",
+    "expo-notifications": "0.32.16",
+    "expo-router": "6.0.22",
+    "expo-secure-store": "15.0.8",
+    "expo-sharing": "14.0.8",
+    "expo-splash-screen": "31.0.13",
     "expo-squircle-view": "1.1.0",
     "expo-standard-web-crypto": "1.8.1",
-    "expo-status-bar": "2.2.3",
-    "expo-system-ui": "5.0.11",
-    "expo-updates": "0.28.17",
+    "expo-status-bar": "3.0.9",
+    "expo-system-ui": "6.0.9",
+    "expo-updates": "29.0.16",
     "fast-text-encoding": "1.0.6",
     "immer": "10.1.1",
     "just-memoize": "2.2.0",
     "lodash.get": "4.4.2",
-    "lottie-react-native": "7.2.2",
+    "lottie-react-native": "7.3.5",
     "mixpanel-react-native": "3.1.2",
     "p-memoize": "7.1.1",
     "punycode": "2.3.1",
-    "react": "19.0.0",
-    "react-dom": "19.0.0",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
     "react-hook-form": "7.62.0",
-    "react-native": "0.79.6",
+    "react-native": "0.81.5",
     "react-native-draggable-flatlist": "4.0.1",
     "react-native-edge-to-edge": "1.6.0",
-    "react-native-gesture-handler": "2.24.0",
+    "react-native-gesture-handler": "2.28.0",
     "react-native-get-random-values": "1.11.0",
     "react-native-keyboard-aware-scroll-view": "0.9.5",
-    "react-native-keyboard-controller": "1.17.3",
+    "react-native-keyboard-controller": "1.18.5",
     "react-native-localize": "3.4.1",
     "react-native-qrcode-svg": "6.3.15",
-    "react-native-reanimated": "4.1.5",
-    "react-native-safe-area-context": "5.4.0",
-    "react-native-screens": "4.11.1",
-    "react-native-svg": "15.11.2",
+    "react-native-reanimated": "4.1.6",
+    "react-native-safe-area-context": "5.6.2",
+    "react-native-screens": "4.16.0",
+    "react-native-svg": "15.12.1",
     "react-native-svg-transformer": "1.3.0",
     "react-native-view-shot": "4.0.3",
-    "react-native-webview": "13.14.1",
+    "react-native-webview": "13.15.0",
     "react-native-worklets": "0.6.1",
     "react-redux": "9.2.0",
     "readable-stream": "4.5.2",
@@ -168,10 +167,10 @@
     "@lingui/format-po": "5.4.0",
     "@lingui/macro": "5.4.0",
     "@testing-library/react": "16.3.0",
-    "@types/jest": "29.5.12",
+    "@types/jest": "29.5.14",
     "@types/node": "24.0.8",
-    "@types/react": "19.0.12",
-    "babel-preset-expo": "13.0.0",
+    "@types/react": "19.1.17",
+    "babel-preset-expo": "54.0.10",
     "concurrently": "8.2.2",
     "happy-dom": "20.0.11",
     "knip": "5.50.5",
