+++ apps/mobile/src/app/(tabs)/(index)/index.tsx
@@ -1,6 +1,6 @@
 import { HomeScreenWithAccount } from '@/components/home/home-with-account-screen';
 import { HomeScreenWithoutAccount } from '@/components/home/home-without-account-screen';
-import { CurrentAccountLoader } from '@/store/settings/settings.read';
+import { CurrentAccountLoader } from '@/store/settings/settings';
 
 export default function HomeScreen() {
   return (
