+++ apps/mobile/src/app/(tabs)/activity.tsx
@@ -1,6 +1,6 @@
 import ActivityScreenWithAccount from '@/components/activity/activity-with-account.screen';
 import ActivityScreenWithoutAccount from '@/components/activity/activity-without-account-screen';
-import { CurrentAccountLoader } from '@/store/settings/settings.read';
+import { CurrentAccountLoader } from '@/store/settings/settings';
 
 export default function ActivityScreen() {
   return (
