+++ apps/mobile/src/features/send/forms/stx/sip10-approver.tsx
@@ -7,7 +7,7 @@ import { useBroadcastStacksTransaction } from '@/queries/stacks/use-broadcast-st
 import { useAccounts } from '@/store/accounts/accounts.read';
 import { useStacksSigners } from '@/store/keychains/stacks/stacks-keychains.read';
 import { assertStacksSigner } from '@/store/keychains/stacks/utils';
-import { useNetworkPreferenceStacksNetwork } from '@/store/settings/settings.read';
+import { useNetworkPreferenceStacksNetwork } from '@/store/settings/settings';
 import { t } from '@lingui/core/macro';
 import { deserializeTransaction } from '@stacks/transactions';
