+++ apps/mobile/src/features/token/bitcoin/stamp.tsx
@@ -1,5 +1,5 @@
 import { TokenDetailsProps } from '@/features/token/types';
-import { imageUnavailableLabel } from '@/features/token/utils/image-unavailable-label';
+import { getImageUnavailableLabel } from '@/features/token/utils/image-unavailable-label';
 
 import { StampAsset } from '@leather.io/models';
 import { CollectibleImage, ImageUnavailable } from '@leather.io/ui/native';
@@ -12,14 +12,14 @@ interface StampProps {
 }
 export function Stamp({ item, height, onPress }: StampProps) {
   if (!item.stampUrl) {
-    return <ImageUnavailable height={height} message={imageUnavailableLabel} />;
+    return <ImageUnavailable height={height} message={getImageUnavailableLabel()} />;
   }
   return (
     <CollectibleImage
       src={item.stampUrl}
       alt={item.stamp.toString()}
       height={height}
-      imageUnavailableLabel={imageUnavailableLabel}
+      imageUnavailableLabel={getImageUnavailableLabel()}
       onPress={onPress ? () => onPress({ assetId: serializeAssetId(getAssetId(item)) }) : undefined}
     />
   );
