+++ apps/mobile/src/features/token/utils/image-unavailable-label.ts
@@ -1,7 +1,5 @@
 import { t } from '@lingui/core/macro';
 
-export const imageUnavailableLabel = getImageUnavailableLabel();
-
 // this is convoluted but avoids hardcoding text in the UI library
 export function getImageUnavailableLabel() {
   return t`Image currently unavailable`;
