+++ apps/mobile/src/i18n/locales/en/messages.po
@@ -108,7 +108,7 @@ msgstr "24 Hour Change"
 msgid "Account {accountIndex}"
 msgstr "Account {accountIndex}"
 
-#: src/store/accounts/accounts.ts:22
+#: src/store/accounts/accounts.read.ts:27
 msgid "Account {displayIndex}"
 msgstr "Account {displayIndex}"
 
@@ -1060,7 +1060,7 @@ msgstr "ID copied to clipboard"
 msgid "If unsafe, your wallet could be drained."
 msgstr "If unsafe, your wallet could be drained."
 
-#: src/features/token/utils/image-unavailable-label.ts:7
+#: src/features/token/utils/image-unavailable-label.ts:5
 msgid "Image currently unavailable"
 msgstr "Image currently unavailable"
