+++ apps/web/package.json
@@ -69,8 +69,8 @@
     "mixpanel-browser": "2.71.1",
     "nprogress": "0.2.0",
     "radix-ui": "1.4.2",
-    "react": "19.0.0",
-    "react-dom": "19.0.0",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
     "react-hook-form": "7.62.0",
     "react-markdown": "10.1.0",
     "react-router": "7.12.0",
@@ -99,8 +99,8 @@
     "@types/d3": "7.4.3",
     "@types/node": "24.0.8",
     "@types/nprogress": "0.2.3",
-    "@types/react": "19.0.12",
-    "@types/react-dom": "19.0.4",
+    "@types/react": "19.1.17",
+    "@types/react-dom": "19.1.11",
     "@types/react-syntax-highlighter": "15.5.13",
     "@types/valid-url": "1.0.7",
     "jsdom": "26.1.0",
