+++ package.json
@@ -49,7 +49,7 @@
     "@testing-library/jest-dom": "6.6.3",
     "@testing-library/react": "16.3.0",
     "@trivago/prettier-plugin-sort-imports": "5.2.2",
-    "@types/jest": "29.5.12",
+    "@types/jest": "29.5.14",
     "@types/node": "24.0.8",
     "@vitest/coverage-v8": "2.1.5",
     "audit-ci": "6.6.1",
@@ -242,8 +242,8 @@
     },
     "patchedDependencies": {
       "pofile@1.1.4": "patches/pofile@1.1.4.patch",
-      "expo-secure-store@14.2.3": "patches/expo-secure-store@14.2.3.patch",
-      "react-native-webview@13.14.1": "patches/react-native-webview@13.14.1.patch"
+      "expo-secure-store@15.0.8": "patches/expo-secure-store@15.0.8.patch",
+      "react-native-webview@13.15.0": "patches/react-native-webview@13.15.0.patch"
     },
     "onlyBuiltDependencies": [
       "better-sqlite3",
