+++ packages/cms/package.json
@@ -42,12 +42,12 @@
     "@sanity/eslint-config-studio": "5.0.2",
     "@sanity/icons": "3.7.4",
     "@sanity/vision": "4.6.1",
-    "@types/react": "19.0.12",
+    "@types/react": "19.1.17",
     "concurrently": "8.2.2",
     "nodemon": "3.1.10",
     "prettier": "3.5.1",
-    "react": "19.0.0",
-    "react-dom": "19.0.0",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
     "styled-components": "6.1.18",
     "tsdown": "0.16.5",
     "typescript": "5.9.3",
