+++ packages/state/package.json
@@ -48,8 +48,8 @@
     "@tanstack/react-query": "5.85.5",
     "bignumber.js": "9.1.2",
     "immer": "10.1.1",
-    "react": "19.0.0",
-    "react-dom": "19.0.0",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
     "remeda": "2.21.2",
     "sbtc": "0.3.2",
     "zod": "4.0.17"
@@ -58,7 +58,7 @@
     "@leather.io/prettier-config": "workspace:*",
     "@leather.io/tsconfig-config": "workspace:*",
     "@testing-library/react": "16.3.0",
-    "@types/react": "19.0.12",
+    "@types/react": "19.1.17",
     "happy-dom": "20.0.11",
     "prettier": "3.5.1",
     "tsdown": "0.16.5",
