+++ packages/ui/package.json
@@ -37,28 +37,28 @@
     "@shopify/flash-list": "2.0.2",
     "@shopify/restyle": "2.4.2",
     "dompurify": "3.2.4",
-    "expo": "53.0.25",
-    "expo-asset": "11.1.7",
-    "expo-blur": "14.1.5",
-    "expo-constants": "17.1.8",
-    "expo-font": "13.3.2",
-    "expo-haptics": "14.1.4",
-    "expo-image": "2.4.1",
-    "expo-linear-gradient": "14.1.5",
-    "expo-splash-screen": "0.30.10",
+    "expo": "54.0.32",
+    "expo-asset": "12.0.12",
+    "expo-blur": "15.0.8",
+    "expo-constants": "18.0.13",
+    "expo-font": "14.0.11",
+    "expo-haptics": "15.0.8",
+    "expo-image": "3.0.11",
+    "expo-linear-gradient": "15.0.8",
+    "expo-splash-screen": "31.0.13",
     "expo-squircle-view": "1.1.0",
     "framer-motion": "12.23.22",
     "prism-react-renderer": "2.4.1",
     "prismjs": "1.30.0",
     "radix-ui": "1.4.2",
-    "react": "19.0.0",
-    "react-dom": "19.0.0",
-    "react-native": "0.79.6",
-    "react-native-gesture-handler": "2.24.0",
-    "react-native-reanimated": "4.1.5",
-    "react-native-safe-area-context": "5.4.0",
-    "react-native-svg": "15.11.2",
-    "react-native-webview": "13.14.1",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
+    "react-native": "0.81.5",
+    "react-native-gesture-handler": "2.28.0",
+    "react-native-reanimated": "4.1.6",
+    "react-native-safe-area-context": "5.6.2",
+    "react-native-svg": "15.12.1",
+    "react-native-webview": "13.15.0",
     "react-native-worklets": "0.6.1",
     "remeda": "2.21.2",
     "use-events": "1.4.2"
@@ -74,7 +74,7 @@
     "@leather.io/tsconfig-config": "workspace:*",
     "@microsoft/api-extractor": "7.55.0",
     "@pandacss/dev": "1.7.1",
-    "@react-native-async-storage/async-storage": "2.1.2",
+    "@react-native-async-storage/async-storage": "2.2.0",
     "@storybook/addon-actions": "8.6.15",
     "@storybook/addon-docs": "8.6.15",
     "@storybook/addon-essentials": "8.6.15",
@@ -97,9 +97,9 @@
     "@svgr/webpack": "8.1.0",
     "@testing-library/react": "16.3.0",
     "@types/prismjs": "1.26.5",
-    "@types/react": "19.0.12",
-    "@types/react-dom": "19.0.4",
-    "babel-preset-expo": "13.0.0",
+    "@types/react": "19.1.17",
+    "@types/react-dom": "19.1.11",
+    "babel-preset-expo": "54.0.10",
     "concurrently": "8.2.2",
     "css-loader": "7.1.2",
     "happy-dom": "20.0.11",
