+++ packages/ui/src/components/collectibles/native/collectible-html.native.tsx
@@ -2,7 +2,8 @@ import { ReactNode, useState } from 'react';
 import { StyleSheet } from 'react-native';
 import { WebView } from 'react-native-webview';
 
-import { Box, TouchableOpacity } from '../../../native';
+import { Box } from '../../box/box.native';
+import { TouchableOpacity } from '../../button/touchable-opacity.native';
 import { CollectibleCard } from './collectible-card.native';
 import { ImageUnavailable } from './image-unavailable.native';
