+++ packages/ui/src/components/collectibles/native/collectible-image.native.tsx
@@ -2,7 +2,8 @@ import { ReactNode, useEffect, useState } from 'react';
 
 import { Image } from 'expo-image';
 
-import { Box, Pressable, PressableProps } from '../../../native';
+import { Box } from '../../box/box.native';
+import { Pressable, PressableProps } from '../../pressable/pressable.native';
 import { ImageUnavailable } from './image-unavailable.native';
 
 export interface CollectibleImageProps extends PressableProps {
