+++ packages/ui/src/components/collectibles/native/image-unavailable.native.tsx
@@ -1,7 +1,8 @@
 import { PropsWithChildren, ReactNode } from 'react';
 
 import { Eye1ClosedIcon } from '../../../icons/eye-1-closed-icon.native';
-import { Box, Text } from '../../../native';
+import { Box } from '../../box/box.native';
+import { Text } from '../../text/text.native';
 import { CollectibleCard } from './collectible-card.native';
 
 interface ImageUnavailableProps extends PropsWithChildren {
