+++ patches/react-native-webview@13.15.0.patch
@@ -1,12 +1,12 @@
 diff --git a/android/src/newarch/com/reactnativecommunity/webview/RNCWebViewManager.java b/android/src/newarch/com/reactnativecommunity/webview/RNCWebViewManager.java
-index aa2223e0afdd226b8e1347cfdf51c43978c79a36..da0e2b669ae90e49646a9e07737550dd743517e1 100644
+index e20ae03bf1187f90d1eddce9203ac023ef00062f..767b6ac6cfbea586e581c47010e30222f0565d86 100644
 --- a/android/src/newarch/com/reactnativecommunity/webview/RNCWebViewManager.java
 +++ b/android/src/newarch/com/reactnativecommunity/webview/RNCWebViewManager.java
 @@ -312,6 +312,12 @@ public class RNCWebViewManager extends ViewGroupManager<RNCWebViewWrapper>
      public void setNewSource(RNCWebViewWrapper view, @Nullable ReadableMap value) {
          mRNCWebViewManagerImpl.setSource(view, value);
      }
-+ 
++   
 +    @Override
 +    @ReactProp(name = "source")
 +    public void setSource(RNCWebViewWrapper view, @Nullable ReadableMap value) {
@@ -16,7 +16,7 @@ index aa2223e0afdd226b8e1347cfdf51c43978c79a36..da0e2b669ae90e49646a9e07737550dd
      @Override
      @ReactProp(name = "textZoom")
 diff --git a/src/RNCWebViewNativeComponent.ts b/src/RNCWebViewNativeComponent.ts
-index 7dd9179f1dba9ec0449d208af388887cfaa0b401..d88e9f1bbe00cbc60a3966b0ee0bf871c1104bff 100644
+index 16b65ace4aaabae5c1bb882d6e4dbfe4d452ef30..f275bfabe6467fdc9224dc8de07b8ffe415af161 100644
 --- a/src/RNCWebViewNativeComponent.ts
 +++ b/src/RNCWebViewNativeComponent.ts
 @@ -291,6 +291,15 @@ export interface NativeProps extends ViewProps {
@@ -34,4 +34,4 @@ index 7dd9179f1dba9ec0449d208af388887cfaa0b401..d88e9f1bbe00cbc60a3966b0ee0bf871
 +  }>;
    userAgent?: string;
    injectedJavaScriptObject?: string;
- }
+   paymentRequestEnabled?: boolean;
